refactor where local park data goes to root of main nav
