+++ src/ParkList/ParkList.tsx
@@ -1,24 +1,19 @@
-import { nationalParks } from '../ParkData'
 import './ParkList.scss'
-import ParkBtn from '../ParkBtn/ParkBtn'
-import { LocalParkData } from '../interfaces'
 
 interface ChoosePark {
   choosePark: (parkCode: string) => void;
 }
 
-type Props = LocalParkData | ChoosePark
+interface ParksOnDisplay {
+   parksOnDisplay: Array<object>
+}
 
-const ParkList: React.FC<Props> = props => {
+type ParkListProps = ChoosePark | ParksOnDisplay
+
+const ParkList: React.FC<ParkListProps> = props => {
   const { choosePark } = props as ChoosePark
-  const parksOnDisplay = nationalParks.map(park => {
-    return <ParkBtn
-      key={park.parkCode}
-      name={park.name}
-      parkCode={park.parkCode}
-      choosePark={choosePark}
-    />
-  })
+  const { parksOnDisplay } = props as ParksOnDisplay
+
   return (
     <section className="park-list">
       <button className="park-button">randomize</button>
